scrapers/pt/requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.50.0
beautifulsoup4>=4.12.0
pdfplumber>=0.11.0
